updating the opcodes
now ldm/ldd/std/in instructions don't touch the flags. this is achieved by the alu enable signal. when alu enable is on the alu functions normally and allows the flags to be updated, when the alu enable is off, the alu functions normally according to the function code but it doesnt let the flags update.